Chore : change page files to components
