+++ eatnfit-FE/src/components/MainExercise/MainExercise.tsx
@@ -1,14 +1,9 @@
 import { Divider } from "antd";
-import {
-  Dashboard,
-  Footer,
-  MainExerciseItems,
-  MainRadioButton,
-} from "../../components";
-import { Container, ItemContainer, StyledTitle } from "../MainFoodPage/styles";
+import { Footer, MainExerciseItems } from "..";
+import { Container, ItemContainer } from "../MainFood/styles";
 import { Space } from "./styles";
 
-export default function MainExercisePage() {
+export default function MainExercise() {
   const exerciseList = [
     {
       title: "오전 공복 줄넘기",
@@ -37,13 +32,6 @@ export default function MainExercisePage() {
   ];
   return (
     <Container>
-      <MainRadioButton defaultValue="exercise" />
-      <StyledTitle level={3}>대시보드</StyledTitle>
-      <Dashboard
-        title={["오늘 / 목표 몸무게", "섭취 칼로리", "소모 칼로리"]}
-        description={["70.0kg / 65.0kg", "1,250kcal", "-200kcal"]}
-        width={90}
-      />
       <Divider />
       <ItemContainer>
         <Space>
