+++ eatnfit-FE/src/components/MainExercise/index.ts
@@ -1,0 +1,1 @@
+export { default as MainExercise } from "./MainExercise";
