+++ eatnfit-FE/src/components/MainFood/index.ts
@@ -1,0 +1,1 @@
+export { default as MainFood } from "./MainFood";
