+++ eatnfit-FE/src/components/MainPlan/index.ts
@@ -1,0 +1,1 @@
+export { default as MainPlan } from "./MainPlan";
