--- eatnfit-FE/src/pages/MainCalendarPage/MainCalendarPage.tsx
@@ -1,11 +1,0 @@
-import { useState } from "react";
-import { MainCalendar } from "../../components";
-import { Wrapper } from "./styles";
-
-export default function MainCalendarPage() {
-  return (
-    <Wrapper>
-      <MainCalendar />
-    </Wrapper>
-  );
-}
