--- eatnfit-FE/src/pages/MainCalendarPage/index.ts
@@ -1,1 +1,0 @@
-export { default as MainCalendarPage } from "./MainCalendarPage";
