--- eatnfit-FE/src/pages/MainCalendarPage/styles.ts
@@ -1,13 +1,0 @@
-import styled from "styled-components";
-
-const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  svg {
-    color: gray;
-    margin: 30px 0;
-  }
-`;
-
-export { Wrapper };
