--- eatnfit-FE/src/pages/MainFoodPage/index.ts
@@ -1,1 +1,0 @@
-export { default as MainFoodPage } from './MainFoodPage';
