+++ eatnfit-FE/src/pages/MainPage/styles.ts
@@ -1,1 +1,11 @@
+import { Typography } from "antd";
 import styled from "styled-components";
+
+const { Title } = Typography;
+
+const Container = styled.div``;
+export const StyledTitle = styled(Title)`
+  text-align: center;
+  color: #00a9ff !important;
+`;
+export { Container };
