--- eatnfit-FE/src/pages/MainPlanPage/index.ts
@@ -1,1 +1,0 @@
-export { default as MainPlanPage } from "./MainPlanPage";
