+++ eatnfit-FE/src/pages/index.ts
@@ -7,8 +7,4 @@ export * from "./SearchExercisePage";
 export * from "./FoodRecordPage";
 export * from "./AddFoodPage";
 export * from "./SearchFoodPage";
-export * from "./MainFoodPage";
-export * from "./MainExercisePage";
-export * from "./MainCalendarPage";
-export * from "./MainPlanPage";
 export * from "./MainPage";
