+++ eatnfit-FE/src/utils/index.ts
@@ -1,1 +1,2 @@
 export { default as InputStatus } from "./inputStatus";
+export { default as getFormatDate } from "./getFormatDate";
